+++ NEWS.md
@@ -1,6 +1,8 @@
 # landscapemetrics 1.3.1
 * Bugfixes
     * Return NA if all cells are NA
+* Improvements
+    * Cleaner handling of warnings
 
 # landscapemetrics 1.3
 * Bugfixes